Refactor: Use environment variable for API base URL
This commit updates the API base URL to use the VITE_API_URL environment variable.
This allows for easier configuration of the API endpoint in different environments.
A default value of 'http://localhost:5000' is used if the environment variable is not set.
Also, adds a timeout to the axios instance.